Improvements from Develop

+++ src/components/map/ProposalsList/ProposalsList.tsx
@@ -2,11 +2,11 @@ import CloseIcon from "@mui/icons-material/Close";
 import DeleteForeverIcon from "@mui/icons-material/DeleteForever";
 import DoneIcon from "@mui/icons-material/Done";
 import GradeIcon from "@mui/icons-material/Grade";
-import { Paper } from "@mui/material";
+import { CircularProgress, Paper } from "@mui/material";
 import { Box } from "@mui/system";
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
-import React, { useCallback } from "react";
+import React, { useCallback, useState } from "react";
 
 import { Editor } from "@/components/Editor";
 
@@ -34,6 +34,7 @@ type ProposalsListProps = {
 };
 
 const ProposalsList = ({ username, ...props }: ProposalsListProps) => {
+  const [isDeleting, setIsDeleting] = useState<boolean>(false);
   const rateProposalClick = useCallback(
     (proposal: any, proposalIdx: any, correct: any, wrong: any, award: any) => {
       return props.rateProposal(
@@ -52,9 +53,16 @@ const ProposalsList = ({ username, ...props }: ProposalsListProps) => {
   );
 
   const deleteProposalClick = useCallback(
-    (proposal: any, proposalIdx: any) =>
-      props.deleteProposal(event, props.proposals, props.setProposals, proposal.id, proposalIdx),
-    // TODO: check dependencies to remove eslint-disable-next-line
+    async (proposal: any, proposalIdx: any) => {
+      let deleteOK = false;
+      deleteOK = window.confirm("You are going to permanently delete this proposal. Are you sure?");
+      if (!deleteOK) return;
+
+      setIsDeleting(true);
+      await props.deleteProposal(event, props.proposals, props.setProposals, proposal.id, proposalIdx);
+      setIsDeleting(false);
+    },
+    // TODO: >heck dependencies to remove eslint-disable-next-line
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [props.deleteProposal, props.proposals]
   );
@@ -105,6 +113,7 @@ const ProposalsList = ({ username, ...props }: ProposalsListProps) => {
                       <ContainedButton
                         title="Click if you find this proposal Unhelpful."
                         onClick={() => rateProposalClick(proposal, proposalIdx, false, true, false)}
+                        disabled={isDeleting}
                       >
                         <Box sx={{ display: "flex", alignItems: "center", gap: "4px", fill: "inherit" }}>
                           <CloseIcon
@@ -118,6 +127,7 @@ const ProposalsList = ({ username, ...props }: ProposalsListProps) => {
                       <ContainedButton
                         title="Click if you find this proposal helpful."
                         onClick={() => rateProposalClick(proposal, proposalIdx, true, false, false)}
+                        disabled={isDeleting}
                       >
                         <Box sx={{ display: "flex", alignItems: "center", gap: "4px", fill: "inherit" }}>
                           <DoneIcon
@@ -145,8 +155,19 @@ const ProposalsList = ({ username, ...props }: ProposalsListProps) => {
                           <span style={{ color: "inherit" }}>{shortenNumber(proposal.awards, 2, false)}</span>
                         </Box>
                       </ContainedButton>
-
-                      {!proposal.accepted && proposal.proposer === username && (
+                      {isDeleting && (
+                        <Box
+                          sx={{
+                            display: "flex",
+                            alignItems: "center",
+                            justifyContent: "center",
+                            width: "50px",
+                          }}
+                        >
+                          <CircularProgress size={24} />
+                        </Box>
+                      )}
+                      {!isDeleting && !proposal.accepted && proposal.proposer === username && (
                         <ContainedButton
                           title={"Delete your proposal"}
                           onClick={() => deleteProposalClick(proposal, proposalIdx)}
